Application de règles Sonar pour augmenter la qualité du code
