+++ php/lib/params.php
@@ -2,7 +2,7 @@
 $fichier = "../conf/params.ini";
 
 // On lit les données dans le fichier ini
-$ini_objet = new ini ();
+$ini_objet = new FichierIni ();
 $ini_objet->m_fichier($fichier);
 
 $_SESSION ['urlSite'] = $ini_objet->m_valeur("urlSite", "general");
